add EnumServiceSpec and Move test
